Support patching an original Cleopatra exe
As extracting the EXE_RESOURCE is a bit of a pain, add support for using
an original (albeit stripped of any DRM wrappers) version of the
Pharaoh.exe from Cleopatra as a base, and apply the patches directly to
that.

If such an executable is not provided, the input is assumed to be the
extracted EXE_RESOURCE.

The various patches applied to PharaohResizer.exe's embedded version are
replicated here in the DoCrudeliosPatches() function: it's an impressive
bit of patching I admit to not fully understanding, so kudos!

This was tested against the GOG executable -- which is the only version
of Cleopatra I have -- and it produced an executable identical to the
embedded one save for GOG's patched website address.