Add Auto-resolution mode
If a resolution is not specified, instead patch in some code to
determine it at runtime from calls to GetSystemMetrics().

This seems to work fine on my machine, but probably needs testing on
others: it could be less reliable at some strange resolutions, or on
multi-monitor setups which are different to mine (i.e., wine with two
1440x900 monitors side-by-side).

This mode can be activated by simply not passing a width and height into
the program.

(I've also added a default output filename of PharaohNew.exe, so that
you only need to pass the path to Pharaoh.exe or the extracted
EXE_RESOURCE to get something which works. This should even -- in theory
-- allow windows users to drag-and-drop Pharaoh.exe onto this binary,
were it compiled for windows, and have a PharaohNew.exe appear.)